backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
sqlmodel==0.0.16
psycopg2-binary==2.9.10

# Authentication & Security
firebase-admin==6.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment & Configuration
python-decouple==3.8
python-dotenv==1.0.0
python-multipart==0.0.6

# Data Validation
pydantic[email]==2.5.0

# Excel Processing
openpyxl==3.1.2
portalocker==2.7.0

# Async Tasks
celery==5.3.4
redis==5.0.1

# AI Integration
google-generativeai==0.3.2
openai==1.3.0

# File Storage
google-cloud-storage==2.10.0
aiofiles==23.2.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
watchfiles==1.1.0

# Performance & Monitoring
prometheus-client==0.19.0 